+++ FLTK/Graph.h
@@ -137,7 +137,7 @@ public:
         if (d) push_back(d);
     }
 
-    ~Vector_ref() { for (int i=0; i<owned.size(); ++i) delete owned[i]; }
+    ~Vector_ref() { for (int i=0; i<(int)owned.size(); ++i) delete owned[i]; }
 
     void push_back(T& s) { v.push_back(&s); }
     void push_back(T* p) { v.push_back(p); owned.push_back(p); }
